requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
groq==0.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.23.6
pydantic==2.7.1
pydantic-core==2.18.2
PyJWT==1.7.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.11.0
urllib3==2.2.1
